+++ app/controllers/users_controller.rb
@@ -1,31 +1,16 @@
 class UsersController < ApplicationController
-  before_action :authorize_request, except: :create
-  before_action :find_user, except: %i[create index]
+  before_action :authorize_request, except: [:create, :index, :show]
+  before_action :set_user, except: %i[create index]
 
-  # GET /users
-  # GET /users.json
   def index
     @users = User.all
     render json: @users, status: :ok
   end
 
-  # GET /users/1
-  # GET /users/1.json
   def show
     render json: @user, status: :ok
   end
 
-  # GET /users/new
-  def new
-    @user = User.new
-  end
-
-  # GET /users/1/edit
-  def edit
-  end
-
-  # POST /users
-  # POST /users.json
   def create
     @user = User.new(user_params)
     if @user.save
@@ -35,13 +20,15 @@ class UsersController < ApplicationController
     end
   end
 
-  # PATCH/PUT /users/1
-  # PATCH/PUT /users/1.json
   def update
-    if @user.update(user_params)
-      render json: @user, status: :ok
+    if @user == @current_user
+      if @user.update(user_params)
+        render json: @user, status: :ok
+      else
+        render json: @user.errors, status: :unprocessable_entity
+      end
     else
-      render json: @user.errors, status: :unprocessable_entity
+      render json: { error: 'unauthorized' }, status: :unauthorized
     end
   end
 
@@ -54,25 +41,26 @@ class UsersController < ApplicationController
     end
   end
 
-  # DELETE /users/1
-  # DELETE /users/1.json
   def destroy
-    destroy_comments(@user.comments)
-    destroy_comments(@user.posts)
-    @user.comments.destroy_all;
-    @user.posts.destroy_all;
-    @user.destroy
-    head :no_content
+    if @user == @current_user
+      destroy_comments(@user.comments)
+      destroy_comments(@user.posts)
+      @user.comments.destroy_all;
+      @user.posts.destroy_all;
+      @user.destroy
+      head :no_content
+    else
+      render json: { error: 'unauthorized' }, status: :unauthorized
+    end
   end
 
   private
-  def find_user
+  def set_user
     @user = User.find_by_username!(params[:_username])
     rescue ActiveRecord::RecordNotFound
       render json: { errors: 'User not found' }, status: :not_found
   end
 
-  # Never trust parameters from the scary internet, only allow the white list through.
   def user_params
     params.permit(:name, :username, :email, :password, :password_confirmation)
   end
